+++ src/components/form/RegisterBusinessForm.jsx
@@ -9,6 +9,7 @@ import { injectIntl } from 'react-intl';
 import Button from '../Button';
 import FormDiv from './FormDiv';
 import LocalizedLink from '../link/LocalizedLink';
+import SwitchCheckbox from './SwitchCheckbox';
 
 import {
   FORM_BILLING_EMAIL,
@@ -28,55 +29,6 @@ const FormContainer = styled.div`
 `;
 
 const FormPage = styled.div``;
-
-const SwitchInputSpan = styled.span`
-  display: inline-block;
-  position: relative;
-  width: 60px;
-  height: 34px;
-
-  input {
-    opacity: 0;
-    width: 0;
-    height: 0;
-
-    &:checked + span {
-      background-color: var(--color-primary);
-    }
-
-    &:checked + span::before {
-      transform: translateX(26px);
-    }
-
-    &:focus + span {
-      box-shadow: 0 0 1px #2196f3;
-    }
-  }
-`;
-
-const SwitchSpan = styled.span`
-  position: absolute;
-  cursor: pointer;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  border-radius: 34px;
-  background-color: var(--color-text-weak);
-  transition: 0.4s;
-
-  &::before {
-    position: absolute;
-    content: '';
-    height: 26px;
-    width: 26px;
-    left: 4px;
-    bottom: 4px;
-    border-radius: 50%;
-    background-color: var(--color-background);
-    transition: 0.4s;
-  }
-`;
 
 const RadioDiv = styled.div`
   margin: 0.5rem 0;
@@ -448,17 +400,13 @@ class RegisterBusinessForm extends Component {
               />
             </FormDiv>
             <FormDiv>
-              <label for="same-billing-address">{i('clientRegisterFormSameBillingAddress')}</label>
-              <SwitchInputSpan onClick={this.handleBillingAddressToggleClick}>
-                <input
-                  key={Math.random()}
-                  type="checkbox"
-                  name="isSameBillingAddress"
-                  id="same-billing-address"
-                  defaultChecked={this.state.isSameBillingAddress}
-                />
-                <SwitchSpan />
-              </SwitchInputSpan>
+              <SwitchCheckbox
+                id="same-billing-address"
+                name="isSameBillingAddress"
+                handleClick={this.handleBillingAddressToggleClick}
+                checked={this.state.isSameBillingAddress}
+                label={i('clientRegisterFormSameBillingAddress')}
+              />
             </FormDiv>
           </FormPage>