+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Timeout for the status messages of the contact form.
 - Validation for the input of the contact form.
 - Validation for the input of the client register forms.
+- Component for checkbox inputs styled as switches.
 
 ### Changed